Allow specifying bot username
The assumed username is 'bot' - if the application installer does not
select this username, then the bot process can get in to a loop spamming
itself with "No results found" messages because it is listening on its
own private channel with the querying user. It needs to understand what
is own username is.